Background cleaning in page ftl
Earlier, the FTL did a clean in the foreground when an erase caused it.
Meaning that the erase would have to wait for the clean and then the
erase would be done. This is not how a smart FTL would handle write
induced erase. It should be done in the backgrond. This also affects the
read latency in a closed model, because large latency for writes imply
less writes imply less erases imply less latency in reads.

Implemented background cleaning and did basic testing. Have to run
tail latency experiments.