Clean in the foreground when out of space
If the SSD runs out of space and there are blocks lined up to be
cleaned, it needs to be done in the foreground while making the write to
wait. Added this feature.

Right now, the simulator is running bery slow, probably because of the
many lists that the FTL has to check in order to serve a write. Need to
improve that.