Fixed bug in page and page_pc FTL
If find_write_location returns without any preferred location, then the
increment_log_write_address function should check whether the current
log_write_address still corresponds to an allocated block before
incrementing its page. Another bug was that in find_write_address, if a
free block is chosen, the ret_address page should be set to 0.

Also improved performance of page_pc FTL by not iterating though
allocated list anymore if a match is found in garbage collector.

Added a condition in page_pc that while cleaning the cleaning_block
should be from another plane (otherwise the purpose is defeated). Right
now, if the check fails, an error is printed but the operation is not
aborted. The hope is that with enough number of planes, this should not
happen. Will need to verify that.