Added cached page FTL and fixed bug
Noticed that I had not added the cached page FTL and was making commits
without the actual file being added. Added that file along with a bug
fix in the same.

The page FTL and the page_cache FTL seem to work for now (although they
are still slow). But there are corner cases which cause problems. Will
work on those.